requirements.txt
youtube-transcript-api>=0.6.0
langchain-groq>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters>=0.2.0
langchain-chroma>=0.1.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
streamlit>=1.28.0
